==> analysis/03b_Phylogenetic_Tree_epulos.Rmd <==
---
title: "Phylogenetic Tree Inspection and Rooting"
author: "Calla Bush St George"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/03_Phylogenetic_Tree_epulos/")
```

# Goals

1. Load the FastTree unrooted tree.
2. Add tree to phyloseq object.
3. Visualize and inspect tree with ggtree.
4. Prune ASVs, if needed.
5. Root our tree.
6. Combine new tree with a phyloseq objet.
7. Save 2 phyloseq objects: 1. Unrooted tree phyloseq object, 
                            2. Rooted tree phyloseq object.

# Setting the Environment

## Set the seed
```{r set-seed}
set.seed(238428)
```

## Load packages
```{r load-packages}
pacman::p_load(tidyverse, phyloseq, ggtree, phytools, 
               install = FALSE)
```

## Load Data files
```{r load-data}
# Preprocessed phyloseq object
load("data/06_PreProcessing_epulos/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Load in the tree!
unrooted_tree <- read.tree("data/07_Phylogenetic_Tree_epulos/ASVs_unrooted.tree")
unrooted_tree
str(unrooted_tree)
```

# Merge Phyloseq
```{r merge-physeq}
ntaxa(raw_preprocessed_physeq)
ntaxa(unrooted_tree)

# Merge the tree with the phyloseq object
unrooted_physeq <-
  merge_phyloseq(raw_preprocessed_physeq, unrooted_tree)

unrooted_physeq
```

```{r plot-tree-unrooted}
# Make a basic tree
kingdom_node_tree <-
  ggtree(unrooted_physeq) +
  # Color by kingdom
  geom_tippoint(mapping = aes(color = Kingdom)) +
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "gray")) +
  # Add title
  labs(title = "Unrooted Tree") + 
  # Move legened to the bottom
  theme(legend.position = "bottom") +
  # Add node labels
  geom_text(aes(label = node), hjust = 0.5, vjust = -0.3, size = 2)
kingdom_node_tree
```

# Midroot Tree

```{r midroot-tree}
# Is the tree rooted? 
new_unrooted_tree <- phy_tree(unrooted_physeq)
is.rooted(new_unrooted_tree)

# Let's midpoint root the tree
midpoint_rooted_tree <-
  midpoint.root(new_unrooted_tree)

# Is the new tree rooted?
is.rooted(midpoint_rooted_tree)

# Assign to a new phyloseq object: merging subsetted phyloseq with the new rooted tree

# 1. Create the phyloseq object
physeq <-
  raw_preprocessed_physeq

# Merge tree with the new physeq_rm456
midroot_physeq <-
  merge_phyloseq(physeq, midpoint_rooted_tree)
midroot_physeq

# Quick Inspection of tree
ggtree(midroot_physeq) + 
  geom_tippoint(mapping = aes(color = Genus))
```

# Save to a new phyloseq object

```{r save-physeq}
# Save both phyloseq objects with our tree object to one .RData file
save(list = c("unrooted_physeq", "midroot_physeq"),
     file = "data/07_Phylogenetic_Tree_epulos/phytree_preprocessed_physeq.RData")
```



# Session Information
```{r session-info}
devtools::session_info()
```
